[tests] Split blocks_test.py
